Refactor: Comment 파일 분리
- article과 comment 디렉토리로 분리
- Comment component를 분리
