+++ src/components/pages/articles/ArticleDetailPage/components/comment/CommentCount.tsx
@@ -1,0 +1,7 @@
+import { NumberProps } from '../../../common/types';
+
+const CommentCount = ({ totalCount }: NumberProps) => {
+  return <div>댓글 {totalCount}개</div>;
+};
+
+export default CommentCount;
